Add benchmark for LineString IsSimple (with circular rings)
